move code to package com.novoda.demo.encryption
