+++ encryption/src/main/AndroidManifest.xml
@@ -1,18 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-      package="com.novoda"
-      android:versionCode="1"
-      android:versionName="1.0">
+<manifest
+  xmlns:android="http://schemas.android.com/apk/res/android">
 
 
-    <application android:icon="@drawable/icon" android:label="@string/app_name" android:theme="@style/Theme.Demo" >
-        <activity android:name=".Encrypt"
-                  android:label="@string/app_name">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
+  <application
+    android:icon="@drawable/icon"
+    android:label="@string/app_name"
+    android:theme="@style/Theme.Demo">
+    <activity
+      android:name="com.novoda.demo.encryption.EncryptionActivity"
+      android:label="@string/app_name">
+      <intent-filter>
+        <action android:name="android.intent.action.MAIN" />
+        <category android:name="android.intent.category.LAUNCHER" />
+      </intent-filter>
+    </activity>
 
-    </application>
+  </application>
 </manifest>
