+++ encryption/src/main/java/com/novoda/demo/encryption/Base64DecoderException.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.novoda;
+package com.novoda.demo.encryption;
 
 /**
  * Exception thrown when encountering an invalid Base64 input character.
