+++ encryption/src/main/res/values/strings.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="hello">Hello World, Encrypt!</string>
-    <string name="app_name">Encrypt</string>
+    <string name="app_name">Encryption</string>
 </resources>
